# api-spec.yaml
openapi: 3.1.0
info:
  title: Median Code API
  description: Production-ready FastAPI code generation and API entity management
    with Clerk JWT authentication.
  version: 1.0.0
paths:
...
  /v1/objects:
    get:
      tags:
      - Objects
      summary: List all objects
      description: Retrieve all object definitions accessible to the authenticated
        user.
      operationId: list_objects_v1_objects_get
      security:
      - HTTPBearer: []
      parameters:
      - name: namespace_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Namespace Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ObjectResponse'
                title: Response List Objects V1 Objects Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Objects
      summary: Create a new object
      description: Create a new object definition.
      operationId: create_object_v1_objects_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObjectCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/objects/{object_id}:
    get:
      tags:
      - Objects
      summary: Get object by ID
      description: Retrieve a specific object by its ID.
      operationId: get_object_v1_objects__object_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: object_id
        in: path
        required: true
        schema:
          type: string
          title: Object Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Objects
      summary: Update object
      description: Update an existing object definition.
      operationId: update_object_v1_objects__object_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: object_id
        in: path
        required: true
        schema:
          type: string
          title: Object Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObjectUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Objects
      summary: Delete object
      description: Delete an object. Cannot delete if used in endpoints.
      operationId: delete_object_v1_objects__object_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: object_id
        in: path
        required: true
        schema:
          type: string
          title: Object Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    DerivedRelationshipResponse:
      properties:
        name:
          type: string
          title: Name
        sourceObject:
          type: string
          title: Sourceobject
        sourceObjectId:
          type: string
          format: uuid
          title: Sourceobjectid
        sourceField:
          type: string
          title: Sourcefield
        kind:
          type: string
          title: Kind
        side:
          type: string
          title: Side
        impliesFk:
          anyOf:
          - type: string
          - type: 'null'
          title: Impliesfk
        junctionTable:
          anyOf:
          - type: string
          - type: 'null'
          title: Junctiontable
        required:
          type: boolean
          title: Required
          default: true
      type: object
      required:
      - name
      - sourceObject
      - sourceObjectId
      - sourceField
      - kind
      - side
      title: DerivedRelationshipResponse
      description: 'Read-only derived (incoming) relationship computed from other
        objects.


        :ivar name: Inverse field name on this object.

        :ivar source_object: Name of the object that authored the relationship.

        :ivar source_object_id: ID of the object that authored the relationship.

        :ivar source_field: Authored relationship field name on the source.

        :ivar kind: Relationship kind.

        :ivar side: Which side of the relationship this object is on.

        :ivar implies_fk: Derived FK column name, or null for many_to_many.

        :ivar junction_table: Junction table name for many_to_many relationships.

        :ivar required: Whether the derived FK column is NOT NULL.'
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ModelValidatorInput:
      properties:
        templateId:
          type: string
          format: uuid
          title: Templateid
        parameters:
          anyOf:
          - additionalProperties:
              type: string
            type: object
          - type: 'null'
          title: Parameters
        fieldMappings:
          additionalProperties:
            type: string
          type: object
          title: Fieldmappings
      type: object
      required:
      - templateId
      - fieldMappings
      title: ModelValidatorInput
      description: 'Request schema for attaching a model validator template to an
        object.


        :ivar template_id: Reference to the model validator template.

        :ivar parameters: Template parameter values.

        :ivar field_mappings: Maps template field mapping keys to actual field names.'
    ModelValidatorResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        templateId:
          type: string
          format: uuid
          title: Templateid
        parameters:
          anyOf:
          - additionalProperties:
              type: string
            type: object
          - type: 'null'
          title: Parameters
        fieldMappings:
          additionalProperties:
            type: string
          type: object
          title: Fieldmappings
      type: object
      required:
      - id
      - templateId
      - fieldMappings
      title: ModelValidatorResponse
      description: 'Response schema for an applied model validator.


        :ivar id: Unique identifier for the applied validator.

        :ivar template_id: Reference to the template.

        :ivar parameters: Template parameter values.

        :ivar field_mappings: Resolved field name mappings.'
...
    ObjectCreate:
      properties:
        namespaceId:
          type: string
          format: uuid
          title: Namespaceid
          examples:
          - 00000000-0000-0000-0000-000000000002
        name:
          type: string
          title: Name
          examples:
          - User
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          examples:
          - User object definition
        members:
          items:
            oneOf:
            - $ref: '#/components/schemas/ScalarMemberInput'
            - $ref: '#/components/schemas/RelationshipMemberInput'
            discriminator:
              propertyName: memberType
              mapping:
                relationship: '#/components/schemas/RelationshipMemberInput'
                scalar: '#/components/schemas/ScalarMemberInput'
          type: array
          title: Members
        validators:
          items:
            $ref: '#/components/schemas/ModelValidatorInput'
          type: array
          title: Validators
      type: object
      required:
      - namespaceId
      - name
      title: ObjectCreate
      description: 'Request schema for creating an object.


        :ivar namespace_id: Namespace this object belongs to.

        :ivar name: Object name.

        :ivar description: Object description.

        :ivar members: Ordered list of scalar and relationship members.

        :ivar validators: Inline model validator definitions for this object.'
    ObjectResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
          examples:
          - 00000000-0000-0000-0007-000000000001
        namespaceId:
          type: string
          format: uuid
          title: Namespaceid
          examples:
          - 00000000-0000-0000-0000-000000000001
        name:
          type: string
          title: Name
          examples:
          - User
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          examples:
          - User account object
        members:
          items:
            oneOf:
            - $ref: '#/components/schemas/ScalarMemberResponse'
            - $ref: '#/components/schemas/RelationshipMemberResponse'
            discriminator:
              propertyName: memberType
              mapping:
                relationship: '#/components/schemas/RelationshipMemberResponse'
                scalar: '#/components/schemas/ScalarMemberResponse'
          type: array
          title: Members
        derivedRelationships:
          items:
            $ref: '#/components/schemas/DerivedRelationshipResponse'
          type: array
          title: Derivedrelationships
        usedInApis:
          items:
            type: string
            format: uuid
          type: array
          title: Usedinapis
          examples:
          - - 00000000-0000-0000-0004-000000000001
        validators:
          items:
            $ref: '#/components/schemas/ModelValidatorResponse'
          type: array
          title: Validators
      type: object
      required:
      - id
      - namespaceId
      - name
      title: ObjectResponse
      description: 'Response schema for object data.


        :ivar id: Unique identifier for the object.

        :ivar namespace_id: Namespace this object belongs to.

        :ivar name: Object name.

        :ivar description: Object description.

        :ivar members: Ordered list of scalar and relationship members.

        :ivar derived_relationships: Incoming relationships from other objects.

        :ivar used_in_apis: Array of API IDs that use this object.

        :ivar validators: Model validators attached to this object.'
    ObjectUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
          examples:
          - UpdatedObjectName
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          examples:
          - Updated description
        members:
          anyOf:
          - items:
              oneOf:
              - $ref: '#/components/schemas/ScalarMemberInput'
              - $ref: '#/components/schemas/RelationshipMemberInput'
              discriminator:
                propertyName: memberType
                mapping:
                  relationship: '#/components/schemas/RelationshipMemberInput'
                  scalar: '#/components/schemas/ScalarMemberInput'
            type: array
          - type: 'null'
          title: Members
        validators:
          anyOf:
          - items:
              $ref: '#/components/schemas/ModelValidatorInput'
            type: array
          - type: 'null'
          title: Validators
      type: object
      title: ObjectUpdate
      description: 'Request schema for updating an object.


        :ivar name: Updated object name.

        :ivar description: Updated description.

        :ivar members: Complete members array for reconcile-by-ID (None = don''t touch).

        :ivar validators: Updated validators (None = don''t touch, [] = clear all).'
...
    RelationshipMemberInput:
      properties:
        id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Id
        memberType:
          type: string
          const: relationship
          title: Membertype
          default: relationship
        name:
          type: string
          title: Name
        targetObjectId:
          type: string
          format: uuid
          title: Targetobjectid
        kind:
          type: string
          enum:
          - one_to_one
          - one_to_many
          - many_to_many
          title: Kind
        inverseName:
          type: string
          title: Inversename
        required:
          type: boolean
          title: Required
          default: true
      type: object
      required:
      - name
      - targetObjectId
      - kind
      - inverseName
      title: RelationshipMemberInput
      description: 'Input schema for a relationship member.


        :ivar id: Existing member ID for reconcile-by-ID updates (omit for new).

        :ivar member_type: Discriminator, always ``"relationship"``.

        :ivar name: Relationship name on the source object.

        :ivar target_object_id: Reference to the target object.

        :ivar kind: Relationship kind.

        :ivar inverse_name: Name for the derived reverse field on the target.

        :ivar required: Whether the derived FK column is NOT NULL.'
    RelationshipMemberResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        memberType:
          type: string
          const: relationship
          title: Membertype
          default: relationship
        name:
          type: string
          title: Name
        targetObjectId:
          type: string
          format: uuid
          title: Targetobjectid
        kind:
          type: string
          title: Kind
        inverseName:
          type: string
          title: Inversename
        required:
          type: boolean
          title: Required
          default: true
      type: object
      required:
      - id
      - name
      - targetObjectId
      - kind
      - inverseName
      title: RelationshipMemberResponse
      description: 'Response schema for a relationship member.


        :ivar id: Unique identifier.

        :ivar member_type: Discriminator, always ``"relationship"``.

        :ivar name: Relationship name.

        :ivar target_object_id: Reference to the target object.

        :ivar kind: Relationship kind.

        :ivar inverse_name: Derived reverse field name.

        :ivar required: Whether the FK column is NOT NULL.'
    ScalarMemberInput:
      properties:
        id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Id
        memberType:
          type: string
          const: scalar
          title: Membertype
          default: scalar
        name:
          type: string
          title: Name
        fieldId:
          type: string
          format: uuid
          title: Fieldid
        role:
          type: string
          enum:
          - pk
          - writable
          - write_only
          - read_only
          - created_timestamp
          - updated_timestamp
          - generated_uuid
          title: Role
          default: writable
        isNullable:
          type: boolean
          title: Isnullable
          default: false
        defaultValue:
          anyOf:
          - type: string
          - type: 'null'
          title: Defaultvalue
      type: object
      required:
      - name
      - fieldId
      title: ScalarMemberInput
      description: 'Input schema for a scalar (field-backed) member.


        :ivar id: Existing member ID for reconcile-by-ID updates (omit for new).

        :ivar member_type: Discriminator, always ``"scalar"``.

        :ivar name: Member name on the object.

        :ivar field_id: Reference to the field definition.

        :ivar role: Structural role of the field.

        :ivar is_nullable: Whether the field is nullable.

        :ivar default_value: Optional literal default value.'
    ScalarMemberResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        memberType:
          type: string
          const: scalar
          title: Membertype
          default: scalar
        name:
          type: string
          title: Name
        fieldId:
          type: string
          format: uuid
          title: Fieldid
        role:
          type: string
          title: Role
        isNullable:
          type: boolean
          title: Isnullable
          default: false
        defaultValue:
          anyOf:
          - type: string
          - type: 'null'
          title: Defaultvalue
      type: object
      required:
      - id
      - name
      - fieldId
      - role
      title: ScalarMemberResponse
      description: 'Response schema for a scalar member.


        :ivar id: Unique identifier.

        :ivar member_type: Discriminator, always ``"scalar"``.

        :ivar name: Member name.

        :ivar field_id: Reference to the field definition.

        :ivar role: Structural role.

        :ivar is_nullable: Whether the field is nullable.

        :ivar default_value: Optional literal default value.'
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer